functions/orchestrator_function/requirements.txt
google-cloud-storage
google-cloud-bigquery
functions-framework